Setup project with instance creation
This project goal is to manage GDKP raids in World of Warcraft classic.

When you create an instance, you can set an amount of gold for an item
that dropped from a boss.

When you closes the instance, you have to set the number of player who was
present to the raid.